build(Dockerfile): - jtb324 - 10/17/25 - 8:46am --- On branch master Your branch is up to date with 'origin/master'.
Changes to be committed:
	new file:   Dockerfile

added a dockerfile for consistent build in all environment:
---
I have heard about (and helped) numerous people having issues installing
the PheWAS package on HPC environments and running into several
different types of errors that can be hard to debug without extensive
knowledge about c++ linking and compilation. This commit adds a docker
image that can install the PheWAS package. It is based on the r-base
image v4.4.0